#1-refactor(frontend): rename services_new to services + add code comments
- Consolidate API layer: move services_new/ → services/
- Add file-level & logic comments to Employer, JobSeeker, Shared pages
- Add comments to component files (EmployerLoginForm, layouts, RouteDecider)
- Improve code documentation for maintainability & onboarding
