+++ it-viec-frontend/src/layout/LayoutCustomer/index.tsx
@@ -2,8 +2,8 @@ import { Outlet, useNavigate } from "react-router-dom";
 import "./LayoutCustomer.scss";
 import { useEffect, useState } from "react";
 import logo from "@/assets/images/nhieu viec (355 x 85 px).png";
-import { logoutApi } from "@/services_new/authApi";
-import { getMyCompanyApi } from "@/services_new/companyApi";
+import { logoutApi } from "@/services/authApi";
+import { getMyCompanyApi } from "@/services/companyApi";
 import MenuItem from "@/components/SiderBar/MenuItem";
 import { FiPieChart } from "react-icons/fi";
 import { MdOutlineSpaceDashboard } from "react-icons/md";
