+++ it-viec-frontend/src/pages/Employer/EmployerRegister/index.tsx
@@ -1,3 +1,5 @@
+// Trang đăng ký tài khoản Employer
+// TODO: chưa implement — đây là placeholder stub, chưa có form đăng ký thực sự
 function EmployerRegister() {
   console.log("EmployerRegister");
