+++ it-viec-frontend/src/pages/Employer/EmployerReports/index.tsx
@@ -1,3 +1,5 @@
+// Trang báo cáo thống kê của Employer
+// Hiện tại chỉ hiển thị tiêu đề, chưa có biểu đồ hay dữ liệu báo cáo
 import EmployerStart from "@/components/EmployerStart";
 import { useTranslation } from "react-i18next";
