+++ it-viec-frontend/src/pages/JobSeeker/CVProfile/index.tsx
@@ -21,8 +21,8 @@ import { Col, DatePicker, Row, Select } from "antd";
 import { Form, Input } from "antd";
 import { FaRegTrashAlt } from "react-icons/fa";
 import { IoCameraOutline } from "react-icons/io5";
-import { updateMyProfileApi } from "@/services_new/seekerApi";
-import { getAllCitiesApi } from "@/services_new/cityApi";
+import { updateMyProfileApi } from "@/services/seekerApi";
+import { getAllCitiesApi } from "@/services/cityApi";
 import { GENDER_OPTIONS, VIETNAM_CITIES } from "@/constants";
 import dayjs from "dayjs";
 import Swal from "sweetalert2";
