+++ it-viec-frontend/src/pages/JobSeeker/JobApplications/index.tsx
@@ -6,10 +6,10 @@ import logo from "@/assets/images/nhieu viec (355 x 85 px).png";
 import { Form, Input, Select } from "antd";
 import ButtonSubmit from "@/components/Button";
 import { VIETNAM_CITIES } from "@/constants";
-import { getJobBySlugApi } from "@/services_new/jobApi";
-import { getMyProfileApi } from "@/services_new/seekerApi";
-import { applyToJobApi } from "@/services_new/applicationApi";
-import { getAllCitiesApi } from "@/services_new/cityApi";
+import { getJobBySlugApi } from "@/services/jobApi";
+import { getMyProfileApi } from "@/services/seekerApi";
+import { applyToJobApi } from "@/services/applicationApi";
+import { getAllCitiesApi } from "@/services/cityApi";
 import Swal from "sweetalert2";
 import { useSeekerStore } from "@/store/seekerStore";
 import { useUserStore } from "@/store/userStore";
