+++ it-viec-frontend/src/pages/JobSeeker/Login/index.tsx
@@ -8,8 +8,8 @@ import ButtonSubmit from "@/components/Button";
 import { useNavigate } from "react-router-dom";
 import Swal from "sweetalert2";
 import type { AuthenticationRequest } from "@/types/request.types";
-import { loginApi } from "@/services_new/authApi";
-import { getMyProfileApi } from "@/services_new/seekerApi";
+import { loginApi } from "@/services/authApi";
+import { getMyProfileApi } from "@/services/seekerApi";
 import { useUserStore } from "@/store/userStore";
 import { useSeekerStore } from "@/store/seekerStore";
 import { useTranslation } from "react-i18next";
