+++ it-viec-frontend/src/services/applicationApi.ts
