+++ it-viec-frontend/src/services/seekerApi.ts
